SCRC_DM/requirements.txt
django==3.2.10
requests
django-admin-tools
django-after-response
django-object-actions
